requirements.txt
streamlit>=1.20
python-dotenv>=1.0.0
openai>=1.30.0

